Revert Graphics.setSize().
There is no cross-platform way to "set the size" of the view. Some platforms
support setting the size, some do not. Those that do not can expose the
mechanism for telling PlayN when the size changed in the event that special
users want to make use of PlayN without the standard view size, but we should
not expose a public, cross platform method that "looks" like it will change the
size of the main game view, because that will just confuse people.

+++ robovm/src/playn/robovm/RoboTextLayout.java
@@ -51,7 +51,7 @@ class RoboTextLayout extends TextLayout {
     NSAttributedStringAttributes attribs = createAttribs(font);
     List<CTLine> lines = wrapLines(new NSAttributedString(text, attribs), wrap.width);
 
-    RoboTextLayout[] layouts = new RoboTextLayout[(int)lines.size()];
+    RoboTextLayout[] layouts = new RoboTextLayout[lines.size()];
     for (int ii = 0; ii < layouts.length; ii++) {
       CTLine line = lines.get(ii);
       CFRange range = line.getStringRange();
